rearrange scene id code
